% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadEnv.R
\name{downloadEnv}
\alias{downloadEnv}
\title{Download Environmental Data}
\usage{
downloadEnv(data, edinfo, fileName = NULL, buffer = c(0, 0, 0))
}
\arguments{
\item{data}{Data containing Longitude, Latitude, and UTC to download matching
environmental data for}

\item{edinfo}{either a edinfo object from \link{getEdinfo} or \link{erddapToEdinfo}
or an ERDDAP dataset ID}

\item{fileName}{name of the file to save downloaded data. If left as the default
\code{NULL}, data will be saved to a temporary folder}

\item{buffer}{numeric vector of the amount to buffer the Longitude, Latitude, and
UTC coordinates by}
}
\value{
if download is successful, invisibly returns the filename. If it fails returns
  \code{FALSE}.
}
\description{
Downloads environmental data matching the coordinates in a set of data
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
