% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erddapList}
\alias{erddapList}
\title{A list of edinfo objects from ERDDAP data sources}
\format{
A list with objects of class edinfo
}
\source{
Southwest Fisheries Science Center / NMFS / NOAA
}
\usage{
erddapList
}
\description{
A list of edinfo objects, mostly used internally for functions. These objects
represent different environmental data sources from ERDDAP servers and are
used to download environmental data.
}
\keyword{datasets}
