% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDateline.R
\name{fixDateline}
\alias{fixDateline}
\title{Fix Longitude Data Across the Dateline for Plotting}
\usage{
fixDateline(data, lonName)
}
\arguments{
\item{data}{data frame with a \code{Longitude} column to adjust for
cross-dateline plotting}

\item{lonName}{name of the Longitude column}
}
\description{
Returns a dataframe identical to the input after checking
   if the \code{Longitude} column crosses the dateline. If it does,
   check whether the center is positive or negative and set all values
   to the same sign.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
