% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdinfo.R
\name{getEdinfo}
\alias{getEdinfo}
\title{Browse a List of Curated Environmental Datasets}
\usage{
getEdinfo()
}
\value{
a list of edinfo list objects
}
\description{
This function gets the list of environmental datasets provided
  as a recommended starting point for various measures
}
\examples{

ediList <- getEdinfo()
ediList[[1]]
ediList[['jplMURSST41']]

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
