% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPXTrack.R
\name{readGPXTrack}
\alias{readGPXTrack}
\title{Read Tracks from a GPX File}
\usage{
readGPXTrack(x)
}
\arguments{
\item{x}{a path to a .gpx file}
}
\value{
a dataframe with columns \code{Latitude}, \code{Longitude},
  \code{UTC}, and \code{Name}
}
\description{
Read in a GPX file and convert the tracks to a dataframe
}
\examples{

gpxFile <- system.file('extdata', 'GPX.gpx', package='PAMmisc')
gpxData <- readGPXTrack(gpxFile)
str(gpxData)

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
