% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelect.R
\name{varSelect}
\alias{varSelect}
\title{Utility for Selecting Variables to Download}
\usage{
varSelect(edinfo)
}
\arguments{
\item{edinfo}{a datalist, either from \link{getEdinfo} or created by
\link{erddapToEdinfo}}
}
\description{
Loops through the available variables in an edinfo object and asks
  whether or not each should be downloaded, then stores the result for
  passing on to \link{formatURL}
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
