% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseEdinfo.R
\name{browseEdinfo}
\alias{browseEdinfo}
\title{Browse a List of Environmental Datasets}
\usage{
browseEdinfo(var = NULL)
}
\arguments{
\item{var}{the name or partial name of a variable to filter the available datasets by}
}
\description{
This function browses the list of selected environmental datasets
  that are recommende as a starting point, and prompts the user to select one to use,
  returning an edinfo object. Also allows user to filter by variable name, matching
  will be attempted using regex
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
