% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimateWavFiles.R
\name{decimateWavFiles}
\alias{decimateWavFiles}
\title{Decimate Wave Files}
\usage{
decimateWavFiles(inDir, outDir, newSr)
}
\arguments{
\item{inDir}{directory of wave files to decimate. Can also be a single .wav file.}

\item{outDir}{directory to write wave files to}

\item{newSr}{sample rate to decimate the files to}
}
\description{
Decimate a folder of .wav files or a single .wav file
  to a new sample rate.
}
\details{
This code is based on R code written by Jay Barlow.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
