% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatURL.R
\name{formatURL}
\alias{formatURL}
\alias{formatURL_erddap}
\alias{formatURL_hycom}
\title{Format URL for Downloading Environmental Data}
\usage{
formatURL(base = NULL, dataset, fileType, vars, source = c("erddap",
  "hycom"), bounds, stride = 1)

formatURL_erddap(base, dataset, fileType = "nc", vars, bounds,
  stride = 1)

formatURL_hycom(base, dataset, fileType = "netcdf", vars, bounds,
  stride = 1)
}
\arguments{
\item{base}{the base URL to download from}

\item{dataset}{the dataset identifier to download}

\item{fileType}{the fileType to download}

\item{vars}{a vector of the names of variables to download}

\item{source}{the source from which the data will be downloaded. This argument
is used for formatting purposes and does not affect the base URL.}

\item{bounds}{a list of the bounds for the data to download. For erddap and
hycom data this should be a list with three vectors containing the min/max
values for time, latitude, and longitude in that order. For example:
\code{list(c(timeMin, timeMax), c(latMin, latMax), c(lonMin, lonMax))}. All
time values must be POSIXct.}

\item{stride}{stride interval for downloading data. A value of 1 (default)
gets every data point, 2 gets every other data point, etc. The same value
is applied to all dimensions.}
}
\description{
This function creates a properly formatted URL for downloading
  environmental data. Users can either provide the base URL, dataset ID, and
  other parameters on their own or select from a list of curated options (see
  ?CHECKMYDATASHITYO for details on this list).
}
\details{

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
