% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncToData.R
\name{ncToData}
\alias{ncToData}
\title{Match Data From a Netcdf File}
\usage{
ncToData(data, nc, buffer = c(0, 0, 0), raw = FALSE, quiet = FALSE)
}
\arguments{
\item{data}{dataframe containing Longitude, Latitude, and UTC to extract matching
variables from the netcdf file}

\item{nc}{name of a netcdf file}

\item{buffer}{vector of Longitude, Latitude, and Time (seconds) to buffer around
each datapoint. All values within the buffer will be used to report the mean,
median, and standard deviation}

\item{raw}{logical flag to return only the raw values of the variables. If \code{TRUE}
the output will be changed to a list with length equal to the number of data points.
Each item in the list will have separate named entries for each variable that will have
all values within the given buffer and all values for any Z coordinates present.}

\item{quiet}{logical flag to show warning messages for possible coordinate mismatch}
}
\value{
original dataframe with three attached columns for each variable in the netcdf
  file, one for each of mean, median, and standard deviation of all values within the buffer
}
\description{
Extracts all variables from a netcdf file matching Longitude,
  Latitude, and UTC coordinates in given dataframe
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
