% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpecAnno.R
\name{readSpecAnno}
\alias{readSpecAnno}
\title{Read Pamguard Spectrogram Annotation Table}
\usage{
readSpecAnno(db, table = "Spectrogram_Annotation")
}
\arguments{
\item{db}{database file to read data from}

\item{table}{name of the Spectrogram Annotation table to read}
}
\value{
a dataframe containing spectrogram annotation data
}
\description{
Reads the Spectrogram Annotation table from a PAMGuard database
  and applies some minor formatting
}
\examples{
\dontrun{
myDb <- 'PamguardDatabase.sqlite3'
specAnno <- readSpecAnno(db)
}
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
