% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightPath.R
\name{straightPath}
\alias{straightPath}
\title{Mark Straight Path Segments in GPS Track}
\usage{
straightPath(gps, nSmall = 10, nLarge = 60, thresh = 10, plot = FALSE)
}
\arguments{
\item{gps}{gps data with columns Heading and UTC (POSIX format)}

\item{nSmall}{number of points to average to get ship's current heading}

\item{nLarge}{number of points to average to get ship's longer trend heading}

\item{thresh}{the amount which \code{nSmall} and \code{nBig} should differ by to
call this a turn}

\item{plot}{logical flag to plot result, \code{gps} must also have columns
Latitude and Longitude}
}
\description{
This function attempts to mark portions of a GPS track where a
  ship is travelling in a straight line by comparing the recent average
  heading with a longer term average heading. If these are different, then the
  ship should be turning. Note this currently does not take in to account time,
  only number of points
}
\details{

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
